Fix markdown switch for front page bodies
This fixes a regression introduced in
e9d0db22167ccaf3bc43d8cf7e2ea759187081b2. The feature was enabled
instead of disabled by default ("false" evaluates to a non-empty string,
which is truthy) and since the front page does not have a `Page`
associated to it, the params for it can’t be set.