layouts/partials: Update header to support Hugo menus
The current stable version allows to define menus, hence it's isn't
needed it anymore to define them in the site parameters dictionary.

This commit changes the header to use the Hugo menus for the general
right side menu and adjust the example site to use the Hugo menus to
show that the header changes work as expected.